requirements.txt
pandas
Django
numpy
psycopg2
psycopg2-binary
beautifulsoup4
lxml
selenium
djangorestframework
python-dotenv